NIFI-7377 Modified plain argument output to use filtering/masking methods in provided utility. Refactored utility methods. Updated unit tests.
