+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-stateless/src/main/java/org/apache/nifi/stateless/runtimes/Program.java
@@ -24,7 +24,6 @@ import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
@@ -47,11 +46,12 @@ public class Program {
     public static final String TOKEN_SEGMENT = "token";
     public static final String ACCESS_SEGMENT = "access";
     public static final String SECRET_SEGMENT = "secret";
-    public static final List<String> SENSITIVE_INDICATORS = Arrays.asList(SENSITIVE_TRUE_JSON_SEGMENT, PASSWORD_SEGMENT, TOKEN_SEGMENT, ACCESS_SEGMENT, SECRET_SEGMENT);
+    public static final String API_KEY_SEGMENT = "api_key";
+    public static final List<String> SENSITIVE_INDICATORS = Arrays.asList(SENSITIVE_TRUE_JSON_SEGMENT, PASSWORD_SEGMENT, TOKEN_SEGMENT, ACCESS_SEGMENT, SECRET_SEGMENT, API_KEY_SEGMENT);
 
-    private static final String JSON_FLAG = "--json";
-    private static final String FILE_FLAG = "--file";
-    private static final String YARN_JSON_FLAG = "--yarnjson";
+    public static final String JSON_FLAG = "--json";
+    public static final String FILE_FLAG = "--file";
+    public static final String YARN_JSON_FLAG = "--yarnjson";
 
     private static boolean isVerbose = true;
 
@@ -86,7 +86,7 @@ public class Program {
         } else if (args[0].equals(RUN_OPENWHISK_ACTION_SERVER) && args.length == 2) {
             runOnOpenWhisk(args, systemClassLoader, narWorkingDirectory);
         } else {
-            System.out.println("Invalid input: " + String.join(",", args));
+            System.out.println("Invalid input: " + formatArgs(args));
             printUsage();
             System.exit(1);
         }
@@ -109,7 +109,7 @@ public class Program {
         } else if (args[5].equals(JSON_FLAG)) {
             json = args[6];
         } else {
-            System.out.println("Invalid input: " + String.join(",", args));
+            System.out.println("Invalid input: " + formatArgs(args));
             printUsage();
             System.exit(1);
             return;
@@ -138,14 +138,14 @@ public class Program {
         } else if (args[2].equals(YARN_JSON_FLAG)) {
             json = args[3].replace(';', ',');
         } else {
-            System.out.println("Invalid input: " + String.join(",", args));
+            System.out.println("Invalid input: " + formatArgs(args));
             printUsage();
             System.exit(1);
             return;
         }
 
         JsonObject jsonObject = new JsonParser().parse(json).getAsJsonObject();
-        System.out.println("Running from json: " + jsonObject.toString());
+        System.out.println("Running from json: " + StatelessSecurityUtility.formatJson(jsonObject));
         final RunnableFlow flow = StatelessFlow.createAndEnqueueFromJSON(jsonObject, systemClassLoader, narWorkingDirectory);
 
         // Run Flow
@@ -185,88 +185,8 @@ public class Program {
      * @return the masked string response
      */
     static String formatArgs(String[] args) {
-        List<String> argsList = new ArrayList<>(Arrays.asList(args));
-        int jsonIndex = determineJsonIndex(argsList);
-
-        if (jsonIndex != -1) {
-            if (isVerbose()) {
-                JsonObject json = new JsonParser().parse(argsList.get(jsonIndex)).getAsJsonObject();
-                String maskedJson = formatJson(json);
-                argsList.add(jsonIndex, maskedJson);
-            } else {
-                argsList.add(jsonIndex, "{...json...}");
-            }
-            argsList.remove(jsonIndex + 1);
-        }
-
-        return String.join(",", argsList);
+        return StatelessSecurityUtility.formatArgs(args, isVerbose());
     }
-
-    /**
-     * Returns a String containing the JSON object with any sensitive values masked.
-     *
-     * @param json the JSON object
-     * @return a masked string
-     */
-    static String formatJson(JsonObject json) {
-        return StatelessSecurityUtility.getLoggableRepresentationOfJsonObject(json);
-    }
-
-    /**
-     * Returns the index of the JSON string in this list (checks {@link #JSON_FLAG} first, then {@link #YARN_JSON_FLAG}). Returns -1 if no JSON is present.
-     *
-     * @param argsList the list of arguments
-     * @return the index of the JSON element or -1
-     */
-    static int determineJsonIndex(List<String> argsList) {
-        int jsonIndex = -1;
-        if (argsList.contains(JSON_FLAG)) {
-            jsonIndex = determineJsonIndex(argsList, JSON_FLAG);
-        } else if (argsList.contains(YARN_JSON_FLAG)) {
-            jsonIndex = determineJsonIndex(argsList, YARN_JSON_FLAG);
-        }
-        return jsonIndex;
-    }
-
-    /**
-     * Returns the index of the JSON string in this list for the given flag. Returns -1 if no JSON is present.
-     *
-     * @param argsList the list of arguments
-     * @param flag     either {@link #JSON_FLAG} or {@link #YARN_JSON_FLAG}
-     * @return the index of the JSON element or -1
-     */
-    static int determineJsonIndex(List<String> argsList, String flag) {
-        // One of the arguments is a JSON string
-        int flagIndex = argsList.indexOf(flag);
-        return flagIndex >= 0 ? flagIndex + 1 : -1;
-    }
-
-    /**
-     * Returns a masked String result given the input if sensitive; the input intact if not.
-     *
-     * @param input the input string
-     * @return masked result if input is sensitive, input otherwise
-     */
-    static String sanitizeString(String input) {
-        if (isSensitive(input)) {
-            return StatelessSecurityUtility.getLoggableRepresentationOfSensitiveValue(input);
-        } else {
-            return input;
-        }
-    }
-
-    /**
-     * Returns {@code true} if the provided {@code input} is determined to be a sensitive value that
-     * needs masking before output. This method uses a series of regular expressions to define common
-     * keywords like {@code secret} or {@code password} that indicate a sensitive value.
-     *
-     * @param input the input string
-     * @return true if the value should be masked
-     */
-    static boolean isSensitive(String input) {
-        return input != null && SENSITIVE_INDICATORS.stream().anyMatch(indicator -> input.toLowerCase().matches(".*" + indicator + ".*"));
-    }
-
 
     private static void printUsage() {
         System.out.println("Usage:");
